+++ src/constants/links.ts
@@ -1,0 +1,12 @@
+/**
+ * Centralized External Links & Resources
+ */
+export const EXTERNAL_LINKS = {
+  article: 'https://taruma.substack.com/p/sceneflow-script-to-screen',
+  articleTitle: 'Introducing SceneFlow: Script-to-Screen Synchronization',
+  kofi: 'https://ko-fi.com/tarumainfo',
+  github: 'https://github.com/taruma/SceneFlow',
+  docs: 'https://github.com/taruma/SceneFlow#readme',
+  changelog: 'https://github.com/taruma/SceneFlow/blob/main/CHANGELOG.md',
+  author: 'https://linktr.ee/tarumainfo',
+} as const;